+++ src/communication/network_sr.c
@@ -39,7 +39,7 @@
 #include "connection_sr.h"
 #include "critical_section.h"
 #include "event_log.h"
-#include "internal_task_worker_pool.hpp"
+#include "internal_tasks_worker_pool.hpp"
 #include "log_impl.h"
 #include "memory_alloc.h"
 #include "message_catalog.h"
@@ -1230,7 +1230,7 @@ net_server_start (const char *server_name)
     }
 
   cubthread::initialize (thread_p);
-  cubthread::internal_tasks_workpool::initialize ();
+  cubthread::internal_tasks_worker_pool::initialize ();
   assert (thread_p == thread_get_thread_entry_info ());
 
 #if defined(WINDOWS)
@@ -1332,7 +1332,7 @@ net_server_start (const char *server_name)
     }
 
   cubthread::finalize ();
-  cubthread::internal_tasks_workpool::finalize ();
+  cubthread::internal_tasks_worker_pool::finalize ();
   er_final (ER_ALL_FINAL);
   csect_finalize_static_critical_sections ();
   (void) sync_finalize_sync_stats ();
